requirements.txt
asgiref==3.7.2
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
